analysis files
These are the analysis files for the paper "Durkheim, indoctrinability and the moral ambivalence of human sociality.  An evolutionary model of cultural transmission, cooperation and group-boundaries."

The zip file "insert_NetLogo_Altruismus_SPATIAL_EXPERIMENT 0.3" includes the NetLogo code for the two simulation models and the R code to read the data from the simulation experiments and to generate the graphs.